refactor: extract shared backend code + add native Windows/macOS GUI roadmap
Prep work for native platform GUI backends (Direct2D on Windows, Core
Graphics on macOS). Extracts duplicated logic from GTK and TUI backends
into shared locations so all future backends reuse the same code:

- Color::to_f32_rgba() in render.rs for D2D/CoreGraphics color conversion
- view_row_to_buf_line/view_row_to_buf_pos_wrap moved from GTK to render.rs
- open_url_in_browser() consolidated in core/engine (was duplicated with
  platform-specific logic in both GTK util.rs and TUI mod.rs)
- Added Native Platform GUIs roadmap section to PLAN.md
