+++ PLAN.md
@@ -25,6 +25,7 @@
 ---
 
 ## Recently Completed
+- **Session 255**: Multi-backend prep for native Windows/macOS GUIs — `Color::to_f32_rgba()` for D2D/CoreGraphics; extracted `view_row_to_buf_line()`/`view_row_to_buf_pos_wrap()` from GTK to shared `render.rs`; consolidated `open_url_in_browser()` in core engine (was duplicated in GTK and TUI with platform-specific logic); added Native Platform GUIs roadmap items.
 - **Session 254**: Windows TUI builds + bug fixes — CI/release for `vcd-windows-x86_64.exe`; picker preview stale chars; insert mode click past EOL; scrollbar drag cursor; git panel discard confirm dialog.
 - **Session 253**: Notification / progress indicator — spinner/bell in per-window status bar for background ops (LSP install, project search/replace); auto-dismiss; click-to-clear; 9 tests.
 - **Session 252**: TUI spell underline bleed fix — `set_cell()`/`set_cell_wide()`/`set_cell_styled()` now reset `modifier` and `underline_color` so spell underlines don't bleed into picker overlays. Added remote editing research item.
@@ -273,6 +274,11 @@
 - [ ] **Announce on Reddit & Hacker News** — "Show HN: VimCode — Vim+VSCode hybrid editor in Rust"; post to r/rust, r/vim, r/neovim. Wait until the editor feels stable enough for first impressions.
 - [ ] **Publish on Flathub** — Submit the Flatpak to Flathub for discoverability and a trusted backlink.
 - [ ] **Additional website pages** — Getting Started guide, screenshots gallery, extension registry browser. Add each to sitemap.
+
+### Native Platform GUIs
+- [ ] **Native Windows GUI (Direct2D + DirectWrite)** — Replace GTK4 on Windows with a native Win32 backend using `windows-rs` + Direct2D rendering + DirectWrite text. Produces `vimcode-win.exe` binary via `win-gui` Cargo feature. Architecture: new `src/win_gui/` directory (~12-18K lines) consuming the same `ScreenLayout` from `render.rs` — zero changes to core engine. DirectWrite provides ClearType subpixel AA (non-negotiable for a code editor). Phase 1: Win32 HWND + D2D render target + DWrite text + keyboard input → single editor pane with syntax highlighting. Phase 2: Tab bar, mouse, completion/hover popups, clipboard. Phase 3: File explorer, native dialogs (IFileOpenDialog), terminal panel, sidebars. Phase 4: Multi-monitor DPI, IME composition, file watching, accessibility. Key files: `src/win_gui/mod.rs` (HWND + message loop), `draw.rs` (D2D rendering), `click.rs` (hit-testing), `input.rs` (WM_KEYDOWN/WM_CHAR/WM_IME → engine key names), `dialogs.rs` (native file dialogs), `util.rs` (DPI, COM, color conversion).
+- [ ] **Native macOS GUI (Core Graphics + Core Text)** — Future: native macOS backend using `objc2` + Core Graphics rendering + Core Text. Would produce a `.app` bundle. Same `ScreenLayout` consumption pattern as Windows backend. Core Text provides native font rendering with proper Retina support. Depends on Windows backend proving the multi-backend pattern works.
+- [ ] **Multi-backend prep work** — Extract shared hit-testing geometry, key binding matching, and open-url logic from GTK/TUI backends into `render.rs` and `core/` so all backends (GTK, TUI, Win-GUI, macOS) reuse the same code. `Color::to_f32_rgba()` for D2D/CoreGraphics. `view_row_to_buf_line()`/`view_row_to_buf_pos_wrap()` in render.rs. `open_url_in_browser()` in core/engine. Consider `notify` crate for cross-platform file watching (replaces GTK's gio::FileMonitor).
 
 ### Remote Editing
 - [ ] **Remote editing over SSH** — Research and design a remote editing story for VimCode. Key questions: (1) **Neovim's approach** — Neovim supports `--remote`, `--server`, and `--headless` modes with a msgpack RPC API; clients connect over stdio/TCP/Unix socket; how much of this is worth emulating? (2) **SSH tunneling** — can VimCode run headless on a remote host with the TUI/GTK frontend on the local machine, forwarding over SSH (à la VS Code Remote SSH)? What's the protocol between frontend and engine? (3) **Headless VimCode** — a `vcd --headless` mode that exposes the engine over a socket/pipe for scripting, testing, or remote frontends; what API surface is needed? (4) **sshfs / FUSE alternative** — simpler approach: open remote files via sshfs mount; what are the LSP/git/terminal implications? (5) **Latency** — how to handle input latency, optimistic rendering, reconnection. Study Neovim's `--headless` + `nvim --listen`, VS Code Remote SSH extension architecture, and Mosh's approach to latency compensation.
