+++ src/gtk/mod.rs
@@ -9659,73 +9659,8 @@ impl App {
     }
 }
 
-/// Map a visible row index (0-based from scroll_top) to the corresponding
-/// buffer line index, skipping lines hidden inside closed folds.
-fn view_row_to_buf_line(
-    view: &crate::core::view::View,
-    scroll_top: usize,
-    view_row: usize,
-    total_lines: usize,
-) -> usize {
-    let mut buf_line = scroll_top;
-    let mut visible = 0usize;
-    while buf_line < total_lines {
-        if view.is_line_hidden(buf_line) {
-            buf_line += 1;
-            continue;
-        }
-        if visible == view_row {
-            return buf_line;
-        }
-        visible += 1;
-        if let Some(fold) = view.fold_at(buf_line) {
-            buf_line = fold.end + 1;
-        } else {
-            buf_line += 1;
-        }
-    }
-    // Clamp to last valid line
-    total_lines.saturating_sub(1)
-}
-
-/// Like `view_row_to_buf_line`, but accounts for word-wrapped lines.
-/// Returns `(buffer_line, segment_col_offset)` — the segment offset is the
-/// character index within the buffer line where the clicked visual segment starts.
-fn view_row_to_buf_pos_wrap(
-    view: &crate::core::view::View,
-    buffer: &crate::core::buffer::Buffer,
-    scroll_top: usize,
-    view_row: usize,
-    total_lines: usize,
-    viewport_cols: usize,
-) -> (usize, usize) {
-    let mut buf_line = scroll_top;
-    let mut visible = 0usize;
-    while buf_line < total_lines {
-        if view.is_line_hidden(buf_line) {
-            buf_line += 1;
-            continue;
-        }
-        // Compute how many visual rows this buffer line occupies when wrapped.
-        let line_str = buffer.content.line(buf_line).to_string();
-        let line_str = line_str.trim_end_matches('\n');
-        let segments = render::compute_word_wrap_segments(line_str, viewport_cols);
-        let visual_rows = segments.len();
-        if view_row < visible + visual_rows {
-            // The clicked row falls within this buffer line.
-            let seg_idx = view_row - visible;
-            let seg_col_offset = segments.get(seg_idx).map(|&(start, _)| start).unwrap_or(0);
-            return (buf_line, seg_col_offset);
-        }
-        visible += visual_rows;
-        if let Some(fold) = view.fold_at(buf_line) {
-            buf_line = fold.end + 1;
-        } else {
-            buf_line += 1;
-        }
-    }
-    (total_lines.saturating_sub(1), 0)
-}
+// view_row_to_buf_line and view_row_to_buf_pos_wrap are now shared functions
+// in render.rs — use render::view_row_to_buf_line / render::view_row_to_buf_pos_wrap.
 
 /// Calculate gutter width in pixels based on line number mode and buffer size
 #[allow(dead_code)]
